Split `HpoDisease` loading into 2 parts to expose HPOA content and `HpoDisease` models.
